requirements.txt
 fastapi==0.115.0
uvicorn[standard]==0.32.0
psycopg[binary,pool]==3.2.3
python-dotenv==1.0.1
pydantic==2.9.2
passlib==1.7.4
bcrypt==4.2.0
pyjwt==2.9.0
python-multipart==0.0.12
